Refactor exception handling by introducing `translate_http_exception` decorator to replace `_translate_http_exception`. Updated relevant endpoints to use the new decorator for improved clarity and maintainability.
