+++ mlflow/gateway/app.py
@@ -1,4 +1,3 @@
-import functools
 from pathlib import Path
 from typing import Any
 
@@ -45,10 +44,13 @@ from mlflow.gateway.constants import (
     MLFLOW_GATEWAY_SEARCH_ROUTES_PAGE_SIZE,
     MLFLOW_QUERY_SUFFIX,
 )
-from mlflow.gateway.exceptions import AIGatewayException
 from mlflow.gateway.providers import get_provider
 from mlflow.gateway.schemas import chat, completions, embeddings
-from mlflow.gateway.utils import SearchRoutesToken, make_streaming_response
+from mlflow.gateway.utils import (
+    SearchRoutesToken,
+    make_streaming_response,
+    translate_http_exception,
+)
 from mlflow.version import VERSION
 
 
@@ -112,24 +114,9 @@ class GatewayAPI(FastAPI):
         return r._to_legacy_route() if (r := self.dynamic_endpoints.get(route_name)) else None
 
 
-def _translate_http_exception(func):
-    """
-    Decorator for translating MLflow exceptions to HTTP exceptions
-    """
-
-    @functools.wraps(func)
-    async def wrapper(*args, **kwargs):
-        try:
-            return await func(*args, **kwargs)
-        except AIGatewayException as e:
-            raise HTTPException(status_code=e.status_code, detail=e.detail)
-
-    return wrapper
-
-
 def _create_chat_endpoint(prov: Provider):
     # https://slowapi.readthedocs.io/en/latest/#limitations-and-known-issues
-    @_translate_http_exception
+    @translate_http_exception
     async def _chat(
         request: Request, payload: chat.RequestPayload
     ) -> chat.ResponsePayload | chat.StreamResponsePayload:
@@ -142,7 +129,7 @@ def _create_chat_endpoint(prov: Provider):
 
 
 def _create_completions_endpoint(prov: Provider):
-    @_translate_http_exception
+    @translate_http_exception
     async def _completions(
         request: Request, payload: completions.RequestPayload
     ) -> completions.ResponsePayload | completions.StreamResponsePayload:
@@ -155,7 +142,7 @@ def _create_completions_endpoint(prov: Provider):
 
 
 def _create_embeddings_endpoint(prov: Provider):
-    @_translate_http_exception
+    @translate_http_exception
     async def _embeddings(
         request: Request, payload: embeddings.RequestPayload
     ) -> embeddings.ResponsePayload:
